proto_tech3-main/src/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pyjwt
passlib
bcrypt==4.0.1
python-multipart
email-validator
fastapi-mail==1.5.0
Authlib==1.6.1
httpx==0.27.2
redis==6.4.0
psycopg2-binary==2.9.10
# sudo apt-get install redis-server